Pushing Version 1.0.1
Here is what this contains.

1. Okay. Translator cannot be broken up into utils right now. it's tightly wound for now.
1. Updated the status codes.
1. ImageAnalysisController has been updated
1. ImageAnalysisUtilities built to support ImageAnalysisController
1. Usual code cleanup and status code updates in the controller.
1. Similar what has been to the other controller, BlobController.
1. Extracted code out of controllers, and started building out utility files so the controller code becomes clean.
    1. ControllerUtilities - utilities to extract away controller actions.
    1. BlobUtilities - utilities to extract away blob creation related actions.
1. Updating the different endpoints to have proper HTTP codes.
    1. Except the Weather Controller. I am leaving it alone for now.
