+++ PhoBloberWebAPI/Controllers/ImageAnalysisController.cs
@@ -10,6 +10,7 @@ using Microsoft.Azure.CognitiveServices.Vision.ComputerVision;
 using System.Net;
 using Microsoft.Azure.CognitiveServices.Vision.ComputerVision.Models;
 using PhoBloberWebAPI.Services;
+using PhoBloberWebAPI.Utilities;
 
 namespace PhoBloberWebAPI.Controllers
 {
@@ -22,7 +23,9 @@ namespace PhoBloberWebAPI.Controllers
         private readonly CVSettingsService _cVSettingsService;
         private readonly StorageSettingsService _storageSettingsService;
         protected ResponseDto _response;
-		public ImageAnalysisController(
+        private readonly ImageAnalysisUtilities imageutil;
+        private readonly ControllerUtilities controlUtil;
+        public ImageAnalysisController(
             IComputerVisionStuff computerVisionStuff,
 			IBlobStorageStuff blobStorageStuff,
             CVSettingsService cVSettingsService,
@@ -34,10 +37,13 @@ namespace PhoBloberWebAPI.Controllers
             this._storageSettingsService = storageSettingsService;
             this._cVSettingsService = cVSettingsService;
 			this._response = new ResponseDto();
-		}
+            imageutil = new ImageAnalysisUtilities();
+            controlUtil = new ControllerUtilities();
+
+        }
 
 		[HttpPost("UploadForPhotoAnalysis")]
-		public async Task<ResponseDto> UploadForPhotoAnalysis(PhotoUploadDTO photoUpload)
+		public async Task<IActionResult> UploadForPhotoAnalysis(PhotoUploadDTO photoUpload)
 		{
 			try
 			{
@@ -62,85 +68,35 @@ namespace PhoBloberWebAPI.Controllers
 				{
 					_response.IsSuccess = false;
 					_response.Message += "Container has no public access. Unable to Upload Image";
-					return _response;
-				}
+					//return _response;
+                    return StatusCode(401, _response);
+                }
 
-				//rename the file.
-				//1. I should not use the name as provided by the user
-				//2. if I don't make any name changes, I will get an error because, 
-				//duplicate files are obviously not allowed. 
-				var uploadFileName = Guid.NewGuid().ToString();
-				//this is the name that will be used for the image creation.
-				var getExtensionOfUploadedImage = System.IO.Path.GetExtension(photoUpload.Image?.FileName);
-				//uploadFileName += photoUpload.Image?.FileName;
-				uploadFileName = uploadFileName + getExtensionOfUploadedImage;
-				BlobClient blobClient = containerClient.GetBlobClient(uploadFileName);
+                var analysed = await imageutil.AnalysePhoto(photoUpload,containerClient,_computerVisionStuff,_cVSettingsService);
 
-				//this also works. preferred, simply way. directly using Azure SDK.
-				var info = await blobClient.UploadAsync(photoUpload.Image?.OpenReadStream());
 
-				PhotoUploadedDTO photoUploadedDTO = new PhotoUploadedDTO();
-				photoUploadedDTO.PhotoName = photoUpload.PhotoName;
-				photoUploadedDTO.AccountName = blobClient.AccountName;
-				photoUploadedDTO.BlobName = blobClient.Name;
-				photoUploadedDTO.BlobContainerName = blobClient.BlobContainerName;
-				photoUploadedDTO.BlobUri = blobClient.Uri;
-				photoUploadedDTO.PhotoDescription = photoUpload.PhotoDescription;
-
-				//okay, image analysis
-				//const string ANALYZE_URL_IMAGE = "https://moderatorsampleimages.blob.core.windows.net/samples/sample16.png";
-				string ANALYZE_URL_IMAGE = blobClient.Uri.ToString();
-				//get the keys
-				var CVKeys = _computerVisionStuff.GetMeComputerVisionSettings(_cVSettingsService);
-
-				// Create a client
-				ComputerVisionClient client =
-				  new ComputerVisionClient(new ApiKeyServiceClientCredentials(CVKeys.VISION_KEY))
-				  { Endpoint = CVKeys.VISION_ENDPOINT };
-
-				List<VisualFeatureTypes?> features = new List<VisualFeatureTypes?>()
-				{
-					VisualFeatureTypes.Tags
-				};
-				// Analyze the URL image 
-				ImageAnalysis results = await client.AnalyzeImageAsync(ANALYZE_URL_IMAGE, visualFeatures: features);
-
-				List<ImageTag> imageTags = new List<ImageTag>();
-
-				// Image tags and their confidence score
-				foreach (var tag in results.Tags)
-				{
-					imageTags.Add(tag);
-				}
-
-				var analysed = new PhotoAnalysedDTO();
-				analysed.imageTags = imageTags;
-				analysed.PhotoName = photoUpload.PhotoName;
-				analysed.AccountName = blobClient.AccountName;
-				analysed.BlobName = blobClient.Name;
-				analysed.BlobContainerName = blobClient.BlobContainerName;
-				analysed.BlobUri = blobClient.Uri;
-				analysed.PhotoDescription = photoUpload.PhotoDescription;
-
-				_response.Result = analysed;
+                _response.Result = analysed;
 				_response.Message = "Photo Analysed Successfully";
+                return StatusCode(200, _response);
 
-			}
+            }
 			catch (Azure.RequestFailedException ex)
 			{
 				_response.IsSuccess = false;
 				_response.Message += ex.Message;
-			}
+                return StatusCode(503, _response);
+            }
 			catch (Exception ex)
 			{
 				_response.IsSuccess = false;
 				_response.Message += ex.Message;
-			}
-			return _response;
+                return StatusCode(500, _response);
+            }
+			//return _response;
 		}
 
         [HttpPost("UploadForOCRAnalysis")]
-        public async Task<ResponseDto> UploadForOCRAnalysis(PhotoUploadDTO photoUpload)
+        public async Task<IActionResult> UploadForOCRAnalysis(PhotoUploadDTO photoUpload)
         {
             try
             {
@@ -165,102 +121,30 @@ namespace PhoBloberWebAPI.Controllers
                 {
                     _response.IsSuccess = false;
                     _response.Message += "Container has no public access. Unable to Upload Image";
-                    return _response;
+                    //return _response;
+                    return StatusCode(401, _response);
                 }
 
-                //rename the file.
-                //1. I should not use the name as provided by the user
-                //2. if I don't make any name changes, I will get an error because, 
-                //duplicate files are obviously not allowed. 
-                var uploadFileName = Guid.NewGuid().ToString();
-                //this is the name that will be used for the image creation.
-                var getExtensionOfUploadedImage = System.IO.Path.GetExtension(photoUpload.Image?.FileName);
-                //uploadFileName += photoUpload.Image?.FileName;
-                uploadFileName = uploadFileName + getExtensionOfUploadedImage;
-                BlobClient blobClient = containerClient.GetBlobClient(uploadFileName);
+                //TODO. Check if file size is more than 4 MB before you do OCR analysis.
 
-                //this also works. preferred, simply way. directly using Azure SDK.
-                var info = await blobClient.UploadAsync(photoUpload.Image?.OpenReadStream());
-
-                PhotoUploadedDTO photoUploadedDTO = new PhotoUploadedDTO();
-                photoUploadedDTO.PhotoName = photoUpload.PhotoName;
-                photoUploadedDTO.AccountName = blobClient.AccountName;
-                photoUploadedDTO.BlobName = blobClient.Name;
-                photoUploadedDTO.BlobContainerName = blobClient.BlobContainerName;
-                photoUploadedDTO.BlobUri = blobClient.Uri;
-                photoUploadedDTO.PhotoDescription = photoUpload.PhotoDescription;
-
-                //okay, ocr analysis
-                //string READ_TEXT_URL_IMAGE = "https://raw.githubusercontent.com/Azure-Samples/cognitive-services-sample-data-files/master/ComputerVision/Images/printed_text.jpg";
-                string READ_TEXT_URL_IMAGE = blobClient.Uri.ToString();
-                //get the keys
-                var CVKeys = _computerVisionStuff.GetMeComputerVisionSettings(_cVSettingsService);
-
-                // Create a client
-                ComputerVisionClient client =
-                  new ComputerVisionClient(new ApiKeyServiceClientCredentials(CVKeys.VISION_KEY))
-                  { Endpoint = CVKeys.VISION_ENDPOINT };
-
-                string urlFile = READ_TEXT_URL_IMAGE;
-
-                // Read text from URL
-                var textHeaders = await client.ReadAsync(urlFile);
-                // After the request, get the operation location (operation ID)
-                string operationLocation = textHeaders.OperationLocation;
-                Thread.Sleep(2000);
-
-                // Retrieve the URI where the extracted text will be stored from the Operation-Location header.
-                // We only need the ID and not the full URL
-                const int numberOfCharsInOperationId = 36;
-                string operationId = operationLocation.Substring(operationLocation.Length - numberOfCharsInOperationId);
-
-                // Extract the text
-                ReadOperationResult results;
-                do
-                {
-                    results = await client.GetReadResultAsync(Guid.Parse(operationId));
-                }
-                while ((results.Status == OperationStatusCodes.Running ||
-                    results.Status == OperationStatusCodes.NotStarted));
-
-                // Collect the found text.
-                var foundlines = new List<string>();
-                var textUrlFileResults = results.AnalyzeResult.ReadResults;
-                foreach (ReadResult page in textUrlFileResults)
-                {
-                    foreach (Line line in page.Lines)
-                    {
-                        //Console.WriteLine(line.Text);
-                        foundlines.Add(line.Text);
-                    }
-                }
-
-                var analysed = new PhotoOCRedDTO();
-                analysed.foundlines = foundlines;
-                analysed.PhotoName = photoUpload.PhotoName;
-                analysed.AccountName = blobClient.AccountName;
-                analysed.BlobName = blobClient.Name;
-                analysed.BlobContainerName = blobClient.BlobContainerName;
-                analysed.BlobUri = blobClient.Uri;
-                analysed.PhotoDescription = photoUpload.PhotoDescription;
+                var analysed = await imageutil.OCRPhoto(photoUpload,containerClient,_computerVisionStuff,_cVSettingsService);
 
                 _response.Result = analysed;
                 _response.Message = "Photo Analysed Successfully";
+                return StatusCode(200, _response);
 
             }
             catch (Azure.RequestFailedException ex)
             {
                 _response.IsSuccess = false;
                 _response.Message += ex.Message;
+                return StatusCode(503, _response);
             }
             catch (Exception ex)
             {
-                _response.IsSuccess = false;
-                _response.Message += ex.Message;
+
+                return StatusCode(500, controlUtil.CreateErrorResponse(ex.Message + "Also, you could check if the image size is more than 4MB. That's one common reason why OCR fails"));
             }
-            return _response;
         }
-
-
     }
 }
