+++ PhoBloberWebAPI/PhoBloberWebAPI.csproj
@@ -4,6 +4,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
+	<Version>1.0.1</Version>
   </PropertyGroup>
 
 	<ItemGroup>
